modifiaction softmax

+++ src/model_predictor.py
@@ -47,7 +47,7 @@ class ModelPredictor:
         self.interpreter.set_tensor(input_details[0]['index'], frame)
         self.interpreter.invoke()
         output_data = self.interpreter.get_tensor(output_details[0]['index'])[0]
-        probabilities = tf.nn.softmax(output_data).numpy()
+        probabilities = tf.nn.softmax(output_data.astype(np.float32)).numpy()
         class_id = np.argmax(probabilities)
         confidence = probabilities[class_id]
         logging.debug(f"output_data -> {output_data}")
